refactor: move DatabaseHelper to Helpers folder
